refactor(web): extract path-extract, diff, file-change modules (CS-6 PR2)
Continue decomposing SessionDetail.tsx by pulling out the next three
groups of pure logic that don't depend on React:

- session-detail/path-extract.ts: file-path detection (looksLikeFilePath,
  shouldTreatAsPathKey), recursive path collection (collectPathsFromValue,
  extractPathsFromToolInput), and display-path formatting (getDisplayPath,
  getDisplayTextWithRelativePaths, getFilePathFromInput, formatTrackedPath)
- session-detail/diff.ts: structured diff construction (buildStructuredDiffFromTexts,
  buildKimiEditDiffBlocks, buildPiEditDiffBlocks, extractEditDiff, createDiffBlock)
- session-detail/file-change.ts: tool classification + per-path summary
  aggregation (classifyToolKind, buildFileChangeSummary,
  buildFileChangeSummaryFromActivity, summarizeFileChangeItems) + shared types
  (FileChangeKind, FileChangeRecord, FileChangeSummary, FileChangeSummaryItem)

Each module gets a focused test file (36 new tests). SessionDetail.tsx
deletes its local copies and imports from the new modules: -460 lines.

Refs CS-6 (multi-PR; PR2 of 5).
